requirements.txt
requests
google-cloud-storage
google-cloud-bigquery
dbt-bigquery==1.7.9
